Restore Binder When Subinfo is Null
Close an edge case where if the Subscription Info for the subId
is null, the Binder that has been cleared isn't restored. This
happens legitimately in cases where the INVALID_SUBSCRIPTION_ID is
passed or the DEFAULT_SUBSCRIPTION_ID is passed to getPhoneNumber.

Due to the structure of the code, permissions checks are called
multiple times if there is no number available from the first
"source". Since that source is "carrier", this is a common case.

Not having cleared the binder then results in a failure when
querying the second "source".

Bug: 317673478
Test: atest FrameworksTelephonyTestCases
Merged-In: I6513c3e90d155846eef9aeffb1fb12b27f4bf9de
Change-Id: I6513c3e90d155846eef9aeffb1fb12b27f4bf9de
(cherry picked from commit c40e9606df036438c95249b894a444e7d305dae9)